[Add]いいね機能、コメント機能非ログイン時閲覧用の記述追加

+++ app/controllers/public/favorites_controller.rb
@@ -4,16 +4,14 @@ class Public::FavoritesController < ApplicationController
   #いいね機能（作成、削除）非同期化
   def create
     @post = Post.find(params[:post_id])
-    favorite = current_user.favorites.new(post_id: @post.id)
-    favorite.save
+    current_user.favorites.find_or_create_by(post_id: @post.id)
     render :favorite
   end
 
   def destroy
     @post = Post.find(params[:post_id])
     favorite = current_user.favorites.find_by(post_id: @post.id)
-    favorite.destroy
+    favorite.destroy if favorite
     render :favorite
   end
-
 end
